Fixed the problem (s).
when adding new values there is no Person object yet. So the add functions need to create a new Person object with the entered values and add it to the observed array that underlies the table.
Because the table is bound to the Model the circle is closed.
Of course the observable array should be mutable ;-( --> in the controller us a mutableListOf instead of a listOf.